fixed some small layout issues and renamed some files. I ended up just using one file, instead of using different files.
