+++ Develop/index.js
@@ -1,12 +1,86 @@
-// TODO: Include packages needed for this application
-const getMarkdown = require('./utils/generateMarkdown')
 const inquirer = require('inquirer')
 const fs = require('fs')
 
-const writeReadme = getMarkdown.generateMarkdown();
 
-// TODO: Create an array of questions for user input
-const questions = inquirer.prompt([
+// TODO: Create a function that returns a license badge based on which license is passed in
+// If there is no license, return an empty string
+
+function renderLicenseBadge(license) {
+    if (license) {
+      return `![License](https://img.shields.io/badge/License-${license}-yellow.svg)`
+    } else {
+      return ""
+    }
+  }
+  
+  // TODO: Create a function that returns the license link
+  // If there is no license, return an empty string
+  function renderLicenseLink(license) {
+    if (license) {
+      return `Click [here](https://opensource.org/license/${license}/) for the license info.`
+    } else {
+      return ""
+    }
+  }
+  
+  // TODO: Create a function that returns the license section of README
+  // If there is no license, return an empty string
+  function renderLicenseSection(license) {
+    if (license) {
+      return `## License
+      
+      ${renderLicenseBadge(license)} ${renderLicenseLink(license)}`
+    } else {
+      return ""
+    }
+  }
+  
+  // TODO: Create a function to generate markdown for README
+  const generateMarkdown = ({title, description, installation, usage, license, contributing, tests, github, email}) => (
+    `# ${title}
+    ${renderLicenseBadge(license)}
+    
+    ## Description
+    
+    ${description}
+    
+    ## Table of Contents
+    
+    [Installation](##installation)
+    [Usage](##usage)
+    [License](##license)
+    [Contributing](##contributing)
+    [Tests](##tests)
+    [Questions](##questions)
+    
+    ## Installation
+    
+    ${installation}
+    
+    ## Usage 
+    
+    ${usage}
+    
+    ${renderLicenseSection(license)}
+    
+    ## Contributing
+    
+    ${contributing}
+    
+    ## Tests
+    
+    ${tests}
+    
+    ## Questions
+  
+    This is the name of my Github profile; ${github}. 
+    Here is the link to my GitHub; [github link](https://github.com/${github})
+  
+    If you would like to reach out to me for any questions, you can communicate with me via my email, ${email}`
+    
+  )
+
+  inquirer.prompt([
     {
         type: 'input',
         name: 'title',
@@ -31,11 +105,11 @@ const questions = inquirer.prompt([
         type: 'list',
         name: 'license',
         message: 'Please select a license for your project. If you do not want a license, hit enter.',
-        choices: ['License-MIT', 'Apache-2.0', 'PDDL', 'Zlib']
+        choices: ['MIT', 'ODC_BY', 'Zlib']
     },
     {
         type: 'input',
-        name: 'contribution',
+        name: 'contributing',
         message: 'Please write the contribution guidelines for your project:',
     },
     {
@@ -55,17 +129,8 @@ const questions = inquirer.prompt([
     }
 ]).then((response) => {
     console.log(response)
+    const markup = generateMarkdown(response)
+    fs.writeFile('./README.md', markup, err => {
+        err ? console.error(err) : console.log("File Successfully Written")
+    });
 });
-
-// TODO: Create a function to write README file
-// This is going to create a file named README.me and the data is the getMarkdown, which is retrieving
-// the markdown we created in the separate js file.
-fs.writeFile('./README.md', writeReadme, err => {
-    err ? console.error(err) : console.log("File Successfully Written")
-});
-
-// TODO: Create a function to initialize app
-// function init() {}
-
-// Function call to initialize app
-// init();
